arm: mediatek: split config into separate options for the board and SoC
The current MediaTek Kconfig layout prevents the addition of new board
targets that require the existing SoC drivers.

By making the SoC drivers depend on new CONFIG_MTxxxx variables, it
becomes possible to select these drivers when adding new boards to the
board selection list. The existing TARGET_MTxxxx are then changed to
select the correct SoC driver targets and can continue to be used as a
generic target.

Signed-off-by: Leith Bade <leith@bade.nz>
Series-to: u-boot
Cc: Chunfeng Yun <chunfeng.yun@mediatek.com>
Cc: Ryder Lee <ryder.lee@mediatek.com>
Cc: Weijie Gao <weijie.gao@mediatek.com>
Cc: Sam Shih <sam.shih@mediatek.com>
Commit-notes:
This is my first U-boot patch so I apologise in advance if I am doing
something wrong.

There seems to be a number of different ways to organise the Kconfig
file to support multiple boards using the same CPU. So picked an approach
that required the least number of changes.

My motivation was to be able to add files under a new
board/bananapi/bpi-r3 directory for the Banana Pi BPI-R3 board so I can
add a .env file to set up the required values for the "Standard Boot".

If this change is accepted I plan to follow up with future patches to
add this .env file.
END